Implement proper distribution of scores with BigDecimal to avoid floating point issues

+++ java-exam-evaluator-main/src/main/java/hu/sed/evaluator/task/ExamEvaluator.java
@@ -167,7 +167,7 @@ public class ExamEvaluator implements Task<Score> {
 
     private String printScoredItem(ScoredItem<?> scoredItem) {
         return String.format("\r %.2f/%.2f - %s%s%s",
-                scoredItem.getScore(), scoredItem.getMaxScore(), scoredItem.identifier(),
+                scoredItem.getScore().doubleValue(), scoredItem.getMaxScore().doubleValue(), scoredItem.identifier(),
                 scoredItem.getUnsuccessfulChecks().isEmpty() ? "" : ", Failed checks: " + scoredItem.getUnsuccessfulChecks(), lineSeparator());
     }
